+++ src/main/java/dev/ftb/mods/ftbstuffnthings/integration/kubejs/FusingMachineRecipeSchema.java
@@ -2,7 +2,8 @@ package dev.ftb.mods.ftbstuffnthings.integration.kubejs;
 
 import dev.ftb.mods.ftbstuffnthings.crafting.EnergyRequirement;
 import dev.latvian.mods.kubejs.recipe.RecipeKey;
-import dev.latvian.mods.kubejs.recipe.component.*;
+import dev.latvian.mods.kubejs.recipe.component.FluidStackComponent;
+import dev.latvian.mods.kubejs.recipe.component.IngredientComponent;
 import dev.latvian.mods.kubejs.recipe.schema.RecipeSchema;
 import net.minecraft.world.item.crafting.Ingredient;
 import net.neoforged.neoforge.fluids.FluidStack;