+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [21.1.18]
+
+### Added
+* Wooden Bowls (vanilla) can now be filled from water-containing tanks, taking 250mB water
+* The Dripper can now pull fluid from a tank above itself
+* Machine GUIs (Water Strainer, Fusing Machine, Super Cooler, Tempered Jar) now work when the block is on a Create: Aeronautics / Sable airship sub-level
+  * Added a soft dependency on Sable
+
+### Fixed
+* Fixed Wooden Basin sometimes causing desyncs when block above it breaks
+
 ## [21.1.17]
 
 ### Changed